Made background into a tileable texture - see desc
Desert ground was turned into a texture, along with the addition of an
unlit material. The material can be attached to a object with a quadmesh
and further scripting would allow for a scalable background by tiling
the texture to automatically repeat. Also ignor BackgroundTiler.cs, I
didn't know how to program the math calculations for said scaling.